관리자 회원, 상품 관리

+++ app/Http/Controllers/Api/Admin/ProductController.php
@@ -2,26 +2,38 @@
 
 namespace App\Http\Controllers\Api\Admin;
 
+use App\Enums\ProductCategory;
 use App\Http\Controllers\Api\ApiController;
-use App\Http\Requests\ProductOptionsRequest;
 use App\Http\Requests\ProductRequest;
 use App\Http\Resources\ProductResource;
+use App\Models\Code;
 use App\Models\Product;
+use App\Models\ProductOption;
 use Illuminate\Http\Request;
+use Spatie\MediaLibrary\MediaCollections\Models\Media;
 
 /**
  * @group 관리자
  */
 class ProductController extends ApiController
 {
+
+    public function init()
+    {
+        $categoryItems = ProductCategory::getItems();
+        $productCategoryItems = Code::getItems(Code::PRODUCT_CATEGORY_ID);
+        return $this->respondSuccessfully(compact(['categoryItems', 'productCategoryItems']));
+    }
+
     /** 목록
      * @subgroup Product(상품)
      * @priority 1
      * @responseFile storage/responses/products.json
      */
-    public function index()
+    public function index(Request $request)
     {
-        $items = Product::query()->latest()->paginate($request->take ?? 30);
+        $filters = (array)json_decode($request->input('search'));
+        $items = Product::search($filters)->latest()->paginate($request->take ?? 30);
         return ProductResource::collection($items);
     }
 
@@ -51,12 +63,12 @@ class ProductController extends ApiController
 
         if ($request->file(Product::IMAGES)) {
             foreach ($request->file(Product::IMAGES) as $file) {
-                $product->addMedia($file['file'])->toMediaCollection('product_images');
+                $product->addMedia($file)->toMediaCollection(Product::IMAGES);
             }
         }
         /*if ($request->file(Product::DESC_IMAGES)) {
             foreach ($request->file(Product::DESC_IMAGES) as $file) {
-                $product->addMedia($file['file])->toMediaCollection(Product::DESC_IMAGES);
+                $product->addMedia($file)->toMediaCollection(Product::DESC_IMAGES);
             }
         }*/
 
@@ -68,19 +80,27 @@ class ProductController extends ApiController
      * @priority 1
      * @responseFile storage/responses/product.json
      */
-    public function update(Request $request, Product $product)
+    public function update(ProductRequest $request, Product $product)
     {
         $data = $request->validated();
-        $product = ($product)->update($data);
+        $product->update($data);
+        $options = $data['options'];
+        //$product->options()->createMany($data['options']);
+        $options = array_map(function ($item, $index) use ($product) {
+            $item['id'] = $item['id'] ?? null;
+            $item['product_id'] = $product->id;
+            return $item;
+        }, $options, array_keys($options));
+        ProductOption::upsert($options, ['id'], ['product_id', 'name', 'price', 'original_price', 'stock_quantity']);
 
         if ($request->file(Product::IMAGES)) {
             foreach ($request->file(Product::IMAGES) as $file) {
-                $product->addMedia($file['file'])->toMediaCollection('product_images');
+                $product->addMedia($file)->toMediaCollection(Product::IMAGES);
             }
         }
         /*if ($request->file(Product::DESC_IMAGES)) {
             foreach ($request->file(Product::DESC_IMAGES) as $file) {
-                $product->addMedia($file['file'])->toMediaCollection(Product::DESC_IMAGES);
+                $product->addMedia($file)->toMediaCollection(Product::DESC_IMAGES);
             }
         }*/
 
@@ -93,9 +113,23 @@ class ProductController extends ApiController
      */
     public function destroy(Product $product)
     {
-        //TODO media 파일삭제는???
+        if ($product->orderProducts()->exists()) {
+            abort(500, '주문된 상품이여서 삭제할 수 없습니다.');
+        }
         $product->delete();
+        $product->clearMediaCollection(Product::IMAGES);
         return $this->respondSuccessfully();
     }
 
+    public function destroyImage(Media $media)
+    {
+        $media->delete();
+        return $this->respondSuccessfully();
+    }
+
+    public function destroyOption(ProductOption $productOption)
+    {
+        $productOption->delete();
+        return $this->respondSuccessfully();
+    }
 }
